Extract DOMTestCase
