+++ packages/@stimulus/test/index.ts
@@ -1,3 +1,4 @@
 export * from "./src/application_test_case"
 export * from "./src/controller_test_case"
+export * from "./src/dom_test_case"
 export * from "./src/test_case"
